backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
email-validator==1.1.3